+++ src/main/java/ru/progwards/tasktracker/service/facade/impl/project/ProjectRemoveService.java
@@ -1,11 +1,12 @@
 package ru.progwards.tasktracker.service.facade.impl.project;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
 import org.springframework.stereotype.Service;
 import ru.progwards.tasktracker.repository.dao.Repository;
 import ru.progwards.tasktracker.repository.entity.ProjectEntity;
+import ru.progwards.tasktracker.service.facade.GetService;
 import ru.progwards.tasktracker.service.facade.RemoveService;
+import ru.progwards.tasktracker.controller.exception.DeletionNotPossibleException;
 import ru.progwards.tasktracker.service.vo.Project;
 
 /**
@@ -21,12 +22,22 @@ public class ProjectRemoveService implements RemoveService<Project> {
     @Autowired
     private Repository<Long, ProjectEntity> repository;
 
+    @Autowired
+    private GetService<Long, Project> projectGetService;
+
     /**
      * метод по удалению проекта
      * @param model бизнес-модель проекта, которую необходимо удалить
      */
     @Override
     public void remove(Project model) {
-        repository.delete(model.getId());
+        Project project = projectGetService.get(model.getId());
+
+        // если спискок задач у проекта пустой, то удаляем проект; если какие-то задачи есть, то выводим исключение
+        if (project.getTasks().size() == 0)
+            repository.delete(model.getId());
+        else
+            throw new DeletionNotPossibleException("Project with id = " + model.getId() +
+                    " have tasks. Delete not possible");
     }
 }
